requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30
sqlalchemy>=2.0
alembic
pydantic-settings>=2.0
python-multipart
python-dotenv
pyjwt
passlib[bcrypt]
redis
celery
langchain-core>=0.3.56,<0.4
langchain-openai>=0.2.0
langchain-chroma>=0.1.4
langchain-text-splitters>=0.3.2
chromadb>=0.5.23,<0.6
pypdf>=5.0
python-docx>=1.1.0
